generating pdf and deleting files

+++ controllers/admin.js
@@ -1,5 +1,5 @@
 const Product = require('../models/product');// import class
-
+const fileHelper = require('../util/file');
 const {validationResult} = require('express-validator/check')
 
 exports.getAddProduct = (req, res, next)=>{
@@ -164,7 +164,8 @@ exports.postEditProduct = (req, res, next) => {
         product.price = updatedPrice;
         product.description = updatedDesc;
         if(image) {
-            product.imageUrl = image.path;
+            fileHelper.deletefile(product.imageUrl);
+            product.imageUrl = image.path;//replace image url
         }
         
         return product.save()
@@ -204,14 +205,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
-        .then(() => {
-            console.log('Destroyed Product');
-            res.redirect('/admin/products');
-        })
-        .catch(err => {
-            const error = new Error(err);
-        error.httpStatusCode = 500;
-        return next(error);
-        });
+    Product.findById(prodId).then(product => {
+        if(!product){
+            return next(new Error('Product not found.'));
+        }
+        fileHelper.deletefile(product.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    }).then(() => {
+        console.log('Destroyed Product');
+        res.redirect('/admin/products');
+    })
+    .catch(err => {
+        const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+    });    
 }
